# cmd/sync/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "health check",
                "operationId": "1",
                "responses": {
                    "200": {
                        "description": "health check response with, db connection, sqlite version,and app version ",
                        "schema": {
                            "$ref": "#/definitions/HealthCheck"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "create a new system user",
                "operationId": "create-user",
                "parameters": [
                    {
                        "description": "payload to ceate new app user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AppNewUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " To create new user",
                        "schema": {
                            "$ref": "#/definitions/AppNewUser"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "AppNewUser": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "dateCreated": {
                    "type": "string"
                },
                "dateUpdated": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "HealthCheck": {
            "type": "object",
            "properties": {
                "database": {
                    "type": "integer"
                },
                "db_version": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "maxProcess": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    }
}